+++ CHANGELOG.md
@@ -4,6 +4,9 @@ CHANGELOG](http://keepachangelog.com/) for how to update this file. This project
 adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Fixed
+- Warnings are now logged all the time. Previously they required the `debug`
+  config option to be enabled.
 
 ## [0.5.1] - 2017-08-22
 ### Fixed