+++ CHANGELOG.md
@@ -10,6 +10,8 @@ adheres to [Semantic Versioning](http://semver.org/).
 - Non-error objects are now reported correctly from `window.onerror` and
   `Honeybadger.wrap`. For instance, `throw('string');` will result in the
   uncaught error being reported with the message "string".
+
+### Added
 - The limited stack info in `window.onerror` is now supplemented if the error
   object (in browsers which support it) does not have a stack property, such as
   when `String` is thrown.